Add ExecutionHistory, UnreadGem and Implement #fetch_unread_gems

+++ db/migrate/20160801234751_create_execution_history.rb
@@ -1,8 +1,8 @@
 class CreateExecutionHistory < ActiveRecord::Migration[5.0]
   def change
     create_table :execution_histories do |t|
-      t.integer :latest_tweet_id, null: false
-      t.integer :start_tweet_id, null: false
+      t.bigserial :latest_tweet_id, null: false
+      t.bigserial :start_tweet_id, null: false
     end
   end
 end
